+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [main]
+
+### Changed
+
+- Proximity calculations now ignore the drones whose current position is equal to
+  their takeoff or landing position. This policy was introduced to exclude drones on
+  the ground from proximity calculations in multi-stage shows. When all the drones
+  are at their takeoff or landing positions, proximity is calculated normally.
+
 ## [2.7.0] - 2025-09-02
 
 ### Added